requirements.txt
boto3==1.24.96
streamlit==1.48.0
google-api-python-client==2.179.0
google-auth-httplib2==0.2.0
dateparser==1.2.2
st_audiorec==0.1.3